pyproject.toml
[project]
name = "iot-curtain-control"
version = "2.0.0"
description = "IoT Curtain Control System with Arduino, MQTT, and Flask"
requires-python = ">=3.9"
dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "pyserial>=3.5",
    "paho-mqtt>=1.6.1",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
]

production = [
    "gunicorn>=21.2.0",
    "supervisor>=4.2.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
